+++ Projects/Claims/Sources/Models/FlowClamSingleItemStepModel.swift
@@ -153,6 +153,16 @@ public struct ClaimFlowItemBrandOptionModel: Codable, Equatable, Hashable {
     let itemTypeId: String
 
     init(
+        displayName: String,
+        itemBrandId: String,
+        itemTypeId: String
+    ) {
+        self.displayName = displayName
+        self.itemBrandId = itemBrandId
+        self.itemTypeId = itemTypeId
+    }
+
+    init(
         with model: OctopusGraphQL.FlowClaimSingleItemStepFragment.AvailableItemBrand
     ) {
         self.displayName = model.displayName
